# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media API",
    "version": "1.0.0",
    "description": "Professional REST API documentation for the Social Media Backend. Supports authentication, posts, comments, stories, notifications, real-time messaging, and more.",
    "contact": {
      "name": "API Support",
      "email": "support@socialmedia.dev"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Environment API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أدخل رمز JWT: Bearer <token>"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/feed/newsfeed": {
      "get": {
        "summary": "Get personalized news feed",
        "tags": [
          "Feed"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of feed posts"
          }
        }
      }
    },
    "/feed/trending": {
      "get": {
        "summary": "Get currently trending posts",
        "tags": [
          "Feed"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "timeframe",
            "schema": {
              "type": "string",
              "enum": [
                "1h",
                "24h",
                "7d",
                "30d"
              ]
            },
            "description": "Time window for trending calculation (default 24h)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Number of results (default 20)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number (default 1)"
          }
        ],
        "responses": {
          "200": {
            "description": "Trending posts"
          }
        }
      }
    },
    "/feed/nearby": {
      "get": {
        "summary": "Get posts submitted from nearby location",
        "tags": [
          "Feed"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "radius",
            "schema": {
              "type": "integer"
            },
            "description": "Search radius in meters (default 5000)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Number of results (default 20)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number (default 1)"
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby posts"
          }
        }
      }
    },
    "/feed/suggested-users": {
      "get": {
        "summary": "Get user suggestions to follow",
        "tags": [
          "Feed"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of suggested users"
          }
        }
      }
    },
    "/feed/suggested-posts": {
      "get": {
        "summary": "Get post suggestions based on interactions",
        "tags": [
          "Feed"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of suggested posts"
          }
        }
      }
    },
...
    "/users/{userId}/followers": {
      "get": {
        "summary": "Get followers of a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of followers"
          }
        }
      }
    },
    "/users/{userId}/following": {
      "get": {
        "summary": "Get users followed by a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of following"
          }
        }
      }
    },
...
  }
}